# server/api/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "Visora AI P1", "version": "0.2.0", "description": "本地开发账号服务与运营后台。写请求必须通过浏览器 Origin 精确白名单验证，缺失或不匹配返回403 FORBIDDEN。写方法包含 POST/PATCH/PUT/DELETE，四者都要求 Origin、X-Visora-Request 与 Content-Type 三道校验。浏览器只调用本 origin 的 /api；开发用户端/Admin统一使用127.0.0.1，不混用localhost。代理保留 Origin、Cookie 和 Set-Cookie，不让浏览器直接请求API_TARGET。项目云存储、平台生成任务与积分消耗结算仍未接入；积分流水中的 consume 类型由后续生成迭代写入，当前没有写入方。" },
  "security": [{ "session": [] }],
  "paths": {
...
    "/api/membership/applications": {
      "post": {
        "operationId": "submitMembershipApplication", "description": "已登录用户提交会员申请，进入 pending 供管理员审批。提交不做去重：同一用户可有多条待审申请，审批时逐条处理。",
        "parameters": [{ "$ref": "#/components/parameters/RequestHeader" }],
        "requestBody": { "$ref": "#/components/requestBodies/ApplicationSubmit" },
        "responses": { "201": { "$ref": "#/components/responses/Application" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/admin/users": {
      "get": {
        "operationId": "listUsers", "description": "仅管理员；默认及精确邮箱筛选均包含active/disabled。createdAt和_id倒序，cursor与筛选条件绑定；前端维护cursor栈提供上一页，更换查询或刷新重置cursor栈。不返回密码、sessionVersion或会话。",
        "parameters": [
          { "$ref": "#/components/parameters/Cursor" },
          { "name": "email", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/UserPage" },
          "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "operationId": "createUser", "description": "仅管理员，创建普通 active 用户；拒绝客户端 role/status 字段。",
        "parameters": [{ "$ref": "#/components/parameters/RequestHeader" }],
        "requestBody": { "$ref": "#/components/requestBodies/Credentials" },
        "responses": { "201": { "$ref": "#/components/responses/User" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/admin/users/{id}": {
      "patch": {
        "operationId": "setUserStatus", "description": "仅管理员可停用/恢复普通用户；管理员账号受保护。恢复不能使旧会话复活。",
        "parameters": [{ "$ref": "#/components/parameters/PathID" }, { "$ref": "#/components/parameters/RequestHeader" }],
        "requestBody": { "$ref": "#/components/requestBodies/StatusChange" },
        "responses": { "200": { "$ref": "#/components/responses/User" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/admin/points/ledger": {
      "get": {
        "operationId": "listPointEntries", "description": "仅管理员；积分流水，createdAt/_id 倒序。kind 为 grant/revoke/topup/allocate/allocation_revoke/consume。团队池流水没有 userId，只有 teamId/teamName；成员分配流水同时有 userId 与 teamId。",
        "parameters": [
          { "$ref": "#/components/parameters/Cursor" },
          { "name": "userId", "in": "query", "schema": { "type": "string" } },
          { "name": "kind", "in": "query", "schema": { "type": "string", "enum": ["grant", "revoke", "topup", "allocate", "allocation_revoke", "consume"] } }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/PointEntryPage" },
          "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
    "/api/admin/accounts": {
      "get": {
        "operationId": "listAdmins", "description": "仅管理员；后台账号（role=admin）列表，createdAt/_id 倒序。",
        "parameters": [{ "$ref": "#/components/parameters/Cursor" }],
        "responses": { "200": { "$ref": "#/components/responses/UserPage" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      },
      "post": {
        "operationId": "createAdminAccount", "description": "仅管理员；创建后台账号。邮箱重复返回 409 EMAIL_EXISTS。",
        "parameters": [{ "$ref": "#/components/parameters/RequestHeader" }],
        "requestBody": { "$ref": "#/components/requestBodies/Credentials" },
        "responses": { "201": { "$ref": "#/components/responses/User" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/admin/accounts/{id}": {
      "patch": {
        "operationId": "setAdminStatus", "description": "仅管理员；停用/启用后台账号。不能停用自己（409 SELF_ACTION），不能停用最后一个启用中的管理员（409 LAST_ADMIN）。停用会递增 sessionVersion 使已签发会话立即失效。",
        "parameters": [{ "$ref": "#/components/parameters/PathID" }, { "$ref": "#/components/parameters/RequestHeader" }],
        "requestBody": { "$ref": "#/components/requestBodies/StatusChange" },
        "responses": { "200": { "$ref": "#/components/responses/User" }, "400": { "$ref": "#/components/responses/Error" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" }, "409": { "$ref": "#/components/responses/Error" }, "413": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/admin/permissions": {
      "get": {
        "operationId": "listPermissions", "description": "仅管理员；只读的服务端授权矩阵。按需求对照 D08，这里返回各接口实际要求的身份，不做可编辑的菜单可见性，也不用隐藏菜单冒充授权。",
        "responses": { "200": { "$ref": "#/components/responses/PermissionPage" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    },
    "/api/admin/stats/overview": {
      "get": {
        "operationId": "overview", "description": "仅管理员；运营看板聚合。只返回有真实数据来源的指标：用户/管理员计数、模型与团队计数、待审申请、有效会员、积分发放与撤销合计、用户积分结存、团队池合计、近 7 日注册序列（UTC 自然日）。没有生成任务与积分消耗的数据源，因此不返回生成次数与消耗指标。",
        "responses": { "200": { "$ref": "#/components/responses/Overview" }, "401": { "$ref": "#/components/responses/Error" }, "403": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } }
      }
    }
  },
  "components": {
    "headers": { "NoStore": { "schema": { "type": "string", "const": "no-store" } } },
    "securitySchemes": { "session": { "type": "apiKey", "in": "cookie", "name": "visora_session" } },
    "parameters": {
      "RequestHeader": { "name": "X-Visora-Request", "in": "header", "required": true, "schema": { "type": "string", "const": "1" } },
      "Cursor": { "name": "cursor", "in": "query", "schema": { "type": "string" } },
      "PathID": { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
    },
    "requestBodies": {
      "Credentials": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["email", "password"], "additionalProperties": false, "properties": { "email": { "type": "string", "format": "email" }, "password": { "type": "string", "writeOnly": true } } } } } },
      "StatusChange": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["status"], "additionalProperties": false, "properties": { "status": { "type": "string", "enum": ["active", "disabled"] } } } } } },
...
      "ApplicationSubmit": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["scene", "reason"], "additionalProperties": false, "properties": { "scene": { "type": "string", "maxLength": 60 }, "reason": { "type": "string", "maxLength": 500 } } } } } },
...
    },
    "schemas": {
      "User": { "type": "object", "required": ["id", "email", "role", "status", "points", "createdAt"], "additionalProperties": false, "properties": { "id": { "type": "string" }, "email": { "type": "string", "format": "email" }, "role": { "type": "string", "enum": ["user", "admin"] }, "status": { "type": "string", "enum": ["active", "disabled"] }, "points": { "type": "integer", "description": "本人积分余额。团队池与团队已分配积分是另外两个独立余额，不并入这里。" }, "createdAt": { "type": "string", "format": "date-time" } } },
      "UserResponse": { "type": "object", "required": ["user"], "properties": { "user": { "$ref": "#/components/schemas/User" } } },
      "PointEntry": { "type": "object", "required": ["id", "scope", "kind", "delta", "balanceAfter", "actorId", "createdAt"], "additionalProperties": false, "properties": { "id": { "type": "string" }, "userId": { "type": "string" }, "userEmail": { "type": "string" }, "teamId": { "type": "string" }, "teamName": { "type": "string" }, "scope": { "type": "string", "enum": ["user", "team_pool", "team_member"] }, "kind": { "type": "string", "enum": ["grant", "revoke", "topup", "allocate", "allocation_revoke", "consume"] }, "delta": { "type": "integer", "description": "有符号变动量，负数为减少。" }, "balanceAfter": { "type": "integer", "description": "该 scope 对应余额变动后的值。" }, "modelName": { "type": "string" }, "note": { "type": "string" }, "actorId": { "type": "string" }, "createdAt": { "type": "string", "format": "date-time" } } },
...
      "MembershipApplication": { "type": "object", "required": ["id", "userId", "userEmail", "scene", "reason", "status", "createdAt"], "additionalProperties": false, "properties": { "id": { "type": "string" }, "userId": { "type": "string" }, "userEmail": { "type": "string" }, "scene": { "type": "string" }, "reason": { "type": "string" }, "status": { "type": "string", "enum": ["pending", "approved", "rejected"] }, "decidedBy": { "type": "string" }, "decidedAt": { "type": "string", "format": "date-time" }, "planCode": { "type": "string" }, "createdAt": { "type": "string", "format": "date-time" } } },
...
      "PermissionRule": { "type": "object", "required": ["group", "action", "endpoint", "userAllowed", "note"], "additionalProperties": false, "properties": { "group": { "type": "string" }, "action": { "type": "string" }, "endpoint": { "type": "string" }, "userAllowed": { "type": "boolean", "description": "普通已登录用户是否可访问；false 表示仅管理员。" }, "note": { "type": "string" } } },
      "DailyCount": { "type": "object", "required": ["date", "count"], "additionalProperties": false, "properties": { "date": { "type": "string", "description": "YYYY-MM-DD，UTC。" }, "count": { "type": "integer" } } },
      "OverviewStats": { "type": "object", "required": ["totalUsers", "activeUsers", "disabledUsers", "adminUsers", "newUsersThisMonth", "totalModels", "activeModels", "totalTeams", "activeTeams", "pendingApplications", "activeMemberships", "issuedPoints", "revokedPoints", "outstandingPoints", "teamPoolPoints", "signups"], "additionalProperties": false, "properties": { "totalUsers": { "type": "integer", "description": "role=user 的账号总数。" }, "activeUsers": { "type": "integer" }, "disabledUsers": { "type": "integer" }, "adminUsers": { "type": "integer" }, "newUsersThisMonth": { "type": "integer", "description": "UTC 自然月内新增普通用户。" }, "totalModels": { "type": "integer" }, "activeModels": { "type": "integer" }, "totalTeams": { "type": "integer" }, "activeTeams": { "type": "integer" }, "pendingApplications": { "type": "integer" }, "activeMemberships": { "type": "integer" }, "issuedPoints": { "type": "integer", "description": "scope=user 且 kind=grant 的累计发放。" }, "revokedPoints": { "type": "integer", "description": "累计撤销，展示为正数。" }, "outstandingPoints": { "type": "integer", "description": "普通用户积分结存合计。" }, "teamPoolPoints": { "type": "integer", "description": "所有团队积分池合计，含已解散团队。" }, "signups": { "type": "array", "description": "近 7 日注册序列，含零值日。", "items": { "$ref": "#/components/schemas/DailyCount" } } } },
      "Error": { "type": "object", "required": ["error"], "properties": { "error": { "type": "object", "required": ["code", "message"], "properties": { "code": { "type": "string", "description": "400 INVALID_INPUT；401 INVALID_CREDENTIALS（登录）/UNAUTHENTICATED（会话）；403 FORBIDDEN（角色/Origin/header/Content-Type）；404 NOT_FOUND/MEMBER_NOT_FOUND/PLAN_NOT_FOUND；405 METHOD_NOT_ALLOWED；409 EMAIL_EXISTS/NAME_EXISTS/USER_PROTECTED/TEAM_INACTIVE/MEMBER_EXISTS/MEMBER_HAS_POINTS/INSUFFICIENT_POINTS/ALREADY_DECIDED/SELF_ACTION/LAST_ADMIN；413 PAYLOAD_TOO_LARGE；503 SERVICE_UNAVAILABLE。", "examples": ["INVALID_INPUT", "INVALID_CREDENTIALS", "UNAUTHENTICATED", "FORBIDDEN", "NOT_FOUND", "MEMBER_NOT_FOUND", "PLAN_NOT_FOUND", "METHOD_NOT_ALLOWED", "EMAIL_EXISTS", "NAME_EXISTS", "USER_PROTECTED", "TEAM_INACTIVE", "MEMBER_EXISTS", "MEMBER_HAS_POINTS", "INSUFFICIENT_POINTS", "ALREADY_DECIDED", "SELF_ACTION", "LAST_ADMIN", "PAYLOAD_TOO_LARGE", "SERVICE_UNAVAILABLE"] }, "message": { "type": "string" } } } } }
    },
    "responses": {
      "User": { "description": "账号信息", "headers": { "Cache-Control": { "$ref": "#/components/headers/NoStore" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } } } },
...
      "Application": { "description": "会员申请", "headers": { "Cache-Control": { "$ref": "#/components/headers/NoStore" } }, "content": { "application/json": { "schema": { "type": "object", "required": ["application"], "properties": { "application": { "$ref": "#/components/schemas/MembershipApplication" } } } } } },
      "UserPage": { "description": "账号分页", "headers": { "Cache-Control": { "$ref": "#/components/headers/NoStore" } }, "content": { "application/json": { "schema": { "type": "object", "required": ["items", "nextCursor"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/User" } }, "nextCursor": { "type": ["string", "null"] } } } } } },
      "PointEntryPage": { "description": "积分流水分页", "headers": { "Cache-Control": { "$ref": "#/components/headers/NoStore" } }, "content": { "application/json": { "schema": { "type": "object", "required": ["items", "nextCursor"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/PointEntry" } }, "nextCursor": { "type": ["string", "null"] } } } } } },
...
      "PermissionPage": { "description": "授权矩阵", "headers": { "Cache-Control": { "$ref": "#/components/headers/NoStore" } }, "content": { "application/json": { "schema": { "type": "object", "required": ["items"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/PermissionRule" } } } } } } },
      "Overview": { "description": "看板聚合", "headers": { "Cache-Control": { "$ref": "#/components/headers/NoStore" } }, "content": { "application/json": { "schema": { "type": "object", "required": ["stats"], "properties": { "stats": { "$ref": "#/components/schemas/OverviewStats" } } } } } },
      "Error": { "description": "受控错误；不回显数据库错误或凭据", "headers": { "Cache-Control": { "$ref": "#/components/headers/NoStore" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    }
  }
}